refactor: split T-Timer types into separate file
Addresses @justandras feedback that RundownPlaylist.ts was growing too large.

- Moved all T-Timer related types to RundownPlaylist/TTimers.ts
- Main file reduced from 341 to 204 lines (40% reduction)
- Backward compatibility maintained via index.ts re-exports
- No behavioral changes
