+++ packages/corelib/src/dataModel/RundownPlaylist/index.ts
@@ -1,0 +1,3 @@
+// Re-export all types from both files to maintain backward compatibility
+export * from './RundownPlaylist.js'
+export * from './TTimers.js'
