+++ packages/corelib/src/playout/__tests__/timings.test.ts
@@ -1,6 +1,6 @@
 import { IBlueprintPieceType } from '@sofie-automation/blueprints-integration'
 import {} from 'type-fest'
-import { RundownHoldState } from '../../dataModel/RundownPlaylist.js'
+import { RundownHoldState } from '../../dataModel/RundownPlaylist/index.js'
 import { literal } from '../../lib.js'
 import { calculatePartTimings, CalculateTimingsPiece, PartCalculatedTimings } from '../timings.js'
