+++ packages/corelib/src/snapshots.ts
@@ -12,7 +12,7 @@ import { DBRundown } from './dataModel/Rundown.js'
 import { RundownBaselineAdLibAction } from './dataModel/RundownBaselineAdLibAction.js'
 import { RundownBaselineAdLibItem } from './dataModel/RundownBaselineAdLibPiece.js'
 import { RundownBaselineObj } from './dataModel/RundownBaselineObj.js'
-import { DBRundownPlaylist } from './dataModel/RundownPlaylist.js'
+import { DBRundownPlaylist } from './dataModel/RundownPlaylist/index.js'
 import { DBSegment } from './dataModel/Segment.js'
 import { SofieIngestDataCacheObj } from './dataModel/SofieIngestDataCache.js'
 import { TimelineComplete } from './dataModel/Timeline.js'
